refactor(cmdk): hook to auto focus input on key press
